requirements.txt
django== 5.0.6
djangorestframework==3.15.1
mysqlclient==2.2.4
numpy==2.0.0
matplotlib==3.9.0
django-cors-headers==4.3.1
whitenoise==6.6.0
django-environ==0.11.2
djoser==2.2.3
argon2-cffi==23.1.0
trimesh==4.4.1
torch==2.3.1
torchvision==0.18.1
h5py==3.11.0
pyglet==1.5.29